rag-pipeline/requirements.txt
fastapi>=0.95.0,<0.116.0
uvicorn>=0.22.0,<0.28.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
torch>=2.0.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-exporter-otlp>=1.19.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
prometheus-client>=0.16.0
faiss-cpu>=1.7.0
transformers>=4.30.0
langchain>=0.0.267
langchain-community>=0.0.1
python-multipart>=0.0.6
python-dotenv>=0.21.0
httpx>=0.24.0
sentence-transformers>=2.2.0 
langchain-huggingface>=0.0.1
accelerate>=0.25.0
bitsandbytes>=0.42.0